pyproject.toml
[project]
name = "condo-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu>=1.13.0",
    "numpy<2",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pip>=25.3",
    "pypdf2>=3.0.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
]

[tool.uv]
required-environments = [
"sys_platform == 'darwin' and platform_machine == 'x86_64'"
]

[tool.uv.workspace]
members = [
    ".",
    ".",
]

[tool.uv.sources]
condo-rag = { workspace = true, editable = true }

[dependency-groups]
dev = [
    "condo-rag",
    "fpdf>=1.7.2",
    "pdfplumber>=0.11.8",
    "pytest>=8.4.2",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
